Added matrix initialization, timing, and pragma for DoAllWork function.
Initialized matrix so that all values start at 0, except for the middle value which is set to 100. Added in print statements to display initial state of array if PRINT_ALL_TIME_STEPS is set to true. Added in values and calculations for the timing and performance calculations. Added additional error checking by verifying all numbers at the end of the matrix add up to 100.