Completed DoAllWork functionality for the vertical partition in a column-major traversal.
Implemented code to divide the square matrix into vertical partitions. The nested for loop was designed to traverse through each partion in 'column-major order', ie. starting from the top of each column moving down vertically to the bottom. The first and last row calculations were implemented outside of the inner for loop to optimize the code descreasing the number of if statement checks needed. Tested code with 100 time steps, 64 elements, and 4 threads and the output was identical to the horizontal-parition.